Consult API after subscription
After this change, the use case will be called
after the first subscription, and the value will
be cached and emitted for every other subscription.

The cached value will be erased if there are no
subscriptions after 2 seconds, so the cache is not erased
on configuration changes.
If a subscription happens after the cache is erased a
new call will be made to the use case

Fixes #1

After this change, the use case will be called
after the first subscription, and the value will be cached
and emitted for every other subscription.

The cached value will be erased if there are no subscriptions. If a
subscription happens after the cache is erased a new call will be made
to the use case